Namma Metro Fair Calculator
The Namma Metro Routing and Pricing System is a project that has been developed to provide an efficient and user-friendly travel experience for passengers. The system is designed to utilize four linear data structures to provide accurate information to passengers. Additionally, the system features a Graphical User Interface (GUI) that allows users to input their boarding and destination points with ease.

The system has been built to provide the shortest route, the number of stops, and the cost of the ticket for the journey on the Green and Purple lines of the Namma Metro. This information is presented to the user in a clear and concise manner, making it easy for passengers to plan their journey.

The Namma Metro Routing and Pricing System is an excellent solution for passengers looking for an efficient and reliable way to navigate the Namma Metro. With its user-friendly interface, accurate information, and clear presentation of ticket prices and route details, this system provides an unparalleled travel experience.